auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.4.0"
description = ""
authors = ["Tomas Sanchez <tosanchez@frba.utn.edu.ar>"]
readme = "README.md"
packages = [{ include = "auth", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { version = "^0.97.0", extras = ["all"] }
motor = "^3.1.2"
motor-types = "^1.0.0b2"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
uvicorn = { version = "^0.22.0", extras = ["standard"] }


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = "tests"
python_files = "test_*.py"
junit_family = "xunit1"
log_cli = false
log_level = "DEBUG"